requirements.txt
Flask
psutil
requests
pyyaml
yara-python
